azure/terraform/native/functionCode/agent_image_embeddings_calculator/requirements.txt
azure-functions==1.18.0
Pillow==10.0.0
